COMPLETED THE TRAINING LOOP FOR THE SPhaseGAN
The training loop for the SPhaseGAN has been completed. First Runs
however indicated that there are major bugs in the training loop that
lead to no learing form either the critic or the generator.

Preliminary assessment idicated that the issues is with the loss
function. The current model uses BCE as loss, which seems to be too
simplistic for the use case.

Another issue is the architecture which might not be able to extract
the correct features. The idea behind the current model was to see
whether this simple model could generate any results or not.